+++ src/Components/App.js
@@ -1,24 +1,11 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import { applyMiddleware, compose, createStore } from 'redux';
 import { Provider } from 'react-redux'
-import { rootReducer, watchLoadData } from '../Redux/Reducks';
-import { logger } from 'redux-logger/src';
-import createSagaMiddleware from 'redux-saga'
 import ListItemContainer from './List/ListItem/ListItemContainer';
 import ListDetailsComponent from './List/ListDetailsComponent';
+import { store } from '../Redux/store'
 
 const App = () => {
-    const sagaMiddleware = createSagaMiddleware();
-    const store = createStore(rootReducer,
-        compose(applyMiddleware(
-            sagaMiddleware,
-            logger
-            ),
-            window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-        ));
-    sagaMiddleware.run(watchLoadData);
-    
     return (
         <Provider store={store}>
             <Router>
@@ -28,6 +15,5 @@ const App = () => {
         </Provider>
     )
 }
-
 
 export default App;
