+++ src/Components/List/ListItem/ListItemContainer.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import { loadData } from '../../../Redux/Reducks';
+import { loadData } from '../../../Redux/re-ducks';
 import ListItemComponent from './ListItemComponent';
 
 const ListItemContainer = () => {
